refactor: extract resumes-tag-expansion.ts from resumes.ts
Move tag expansion search types and helpers to
lib/resumes_tag_expansion.ts:
- Types: TagExpansionKeywordGroup, MatchSource, SearchProvenance
- Functions: normalizeTagExpansionKeywordGroups, dedupeProvenance,
  buildTagExpansionSearchQuery, matchesTagExpansionSearchText,
  collectSearchTextProvenance, collectExpandedTerms,
  selectTagExpansionAnchorGroup

resumes.ts re-exports all moved symbols for backward compatibility.
Combined with prior diagnostics extraction, resumes.ts reduced from
3365 → 3097 lines (8% total reduction).
